feat: Dynamically select models
SetActiveModel command added to the language. This allows for
dynamically selecting models as the table is generated, facilitating
the creation of multi-panel tables. When using, one must specify the
FULL set of models in the Vandelay table DHALL file. Then, use the
set_active_models command to pick the SUBSET that should be active for
the following Vandelay commands.